#118: better naming
